move ceil_log2 tests into math module
